+++ src/put/put_var/dictionnary.js
@@ -22,15 +22,12 @@ function encode (value) {
   let bufPos = 8
   for (let i in value) {
     if (value.hasOwnProperty(i)) {
-      value[i].value.copy(buf, bufPos)
+      value[i].copy(buf, bufPos)
       bufPos += value[i].length
     }
   }
 
-  return Promise.resolve({
-    value: buf,
-    length: buf.length
-  })
+  return Promise.resolve(buf)
 }
 
 module.exports = {