+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## CHANGELOG
 
+* v2.1.3
+  * put_var now return only buffer (in promise) 
+
+* v2.1.2
+  * some bugfix
+
+* v2.1.1
+  * int64
+
+* v2.1.0
+  * promisify function
+
 * v2.0.0
   * rethink of the architecture
   * add put / get method like in StreamPeer